requirements.txt
numpy==1.13.3
tensorflow==2.2.0
tensorflow-addons==0.10.0
opencv_contrib_python==4.4.0.44
Pillow==8.3.2
PyYAML==5.4.1
scikit_learn==0.24.2
skimage==0.0